backend/requirements.txt
Flask
Flask-Cors
pandas
elasticsearch
python-dotenv
requests
gunicorn